requirements.txt
beautifulsoup4==4.9.1
Flask==1.1.2
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-language==1.3.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
lxml==4.5.2
pandas==1.1.2
requests==2.24.0